// src/app/core/models/ViewModel/maintenance-offers-details-vm.ts
export class MaintenanceOffersDetailsVM {
    id: any;
    maintenanceOfferId: any;
    productId: any;
    quantity: any;
    price: any;
    valueBeforeTax: any;
    taxValue: any;
    taxRatio: any;
    valueAfterTax: any;
    productNameAr!: string;
    productNameEn!: string;
    maintenancePurchaseBillId:any;
    billed!:boolean;
    purchaseOrderId:any;
    userId: any;
}